requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
requests==2.26.0
beautifulsoup4==4.9.3
pandas==1.3.3
python-dotenv==0.19.0
selenium==4.15.2
mysql-connector-python==8.2.0
numpy==1.26.2
plotly==5.18.0
dash==2.14.1
python-jose==3.3.0
bcrypt==4.0.1 